requirements.txt
# Core dependencies
fastapi[full]==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
python-multipart==0.0.9

# Image processing
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
onnxruntime==1.17.0

# Utilities
rich==13.7.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Development dependencies
black==24.1.1
isort==5.13.2
flake8==7.0.0